Rename folder 'db' to 'storage'
