+++ contracting/stdlib/bridge/imports.py
@@ -1,8 +1,8 @@
 import importlib
 from types import FunctionType, ModuleType
 from contracting.config import PRIVATE_METHOD_PREFIX
-from contracting.db.orm import Datum
-from contracting.db.driver import Driver, OWNER_KEY
+from contracting.storage.orm import Datum
+from contracting.storage.driver import Driver, OWNER_KEY
 from contracting.execution.runtime import rt
 import sys
