+++ contracting/storage/orm.py
@@ -1,4 +1,4 @@
-from contracting.db.driver import Driver
+from contracting.storage.driver import Driver
 from contracting.execution.runtime import rt
 from contracting import config
 from contracting.stdlib.bridge.decimal import ContractingDecimal
